requirements.txt
Flask==2.2.2
Werkzeug==2.2.2
requests
google-cloud-compute
